chore: Bring back tests in concerned files
Part of #2

Some cleanup before continue to work on features

*No functional changes expected*
